Path finding works for one color
FINALLY got the optimal path.

User clicks a start point and an endpoint. The program draws the best path between the two in yellow. If the user clicks the end point again, the path turns black. This is set up to be changeable based on who's turn it is. I still want to figure out how I can make the line flash or make it dotted if it's still not finalized. We'll see.